backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
PyJWT==2.8.0
Werkzeug==3.0.3
